refactor more logic out of the guest crate
