+++ contracts/src/ImageID.sol
@@ -20,5 +20,5 @@ pragma solidity ^0.8.20;
 
 library ImageID {
     bytes32 public constant LIGHT_CLIENT_GUEST_ID =
-        bytes32(0xfbbeefacc250803eb062909a6c6518edd499efac418b135819af0dc8e24e49e9);
+        bytes32(0xba80319e408b161e6b046901affb3c66fb2294b0a87cf88a17cd5c01e21a7850);
 }
