+++ core/src/range_iterator.rs
@@ -15,11 +15,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use blobstream0_primitives::{
+    light_client_verify,
     proto::{TrustedLightBlock, UntrustedLightBlock},
-    LightBlockProveData, DEFAULT_PROVER_OPTS,
+    LightBlockProveData,
 };
-use std::time::Duration;
-use tendermint_light_client_verifier::{types::Header, ProdVerifier, Verdict, Verifier};
+use tendermint_light_client_verifier::{types::Header, Verdict};
 use tendermint_rpc::HttpClient;
 
 use crate::{fetch_untrusted_light_block, fetch_validators};
@@ -100,12 +100,6 @@ impl LightBlockRangeIterator<'_> {
 }
 
 fn validator_stake_overlap(trusted: &TrustedLightBlock, target: &UntrustedLightBlock) -> bool {
-    let vp = ProdVerifier::default();
-    let trusted_state = trusted.as_trusted_state();
-    let target_state = target.as_untrusted_state();
-
     // Replicate same validation as done in the guest, to avoid any inconsistencies
-    let time = (target_state.signed_header.header().time + Duration::from_secs(1)).unwrap();
-    let verdict = vp.verify_update_header(target_state, trusted_state, &DEFAULT_PROVER_OPTS, time);
-    matches!(verdict, Verdict::Success)
+    matches!(light_client_verify(trusted, target), Verdict::Success)
 }
