+++ light-client-guest/guest/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2021"
 risc0-zkvm = { version = "=1.0.5", default-features = false, features = ['std'] }
 blobstream0-primitives = { path = "../../primitives" }
 alloy-sol-types = "0.7"
-alloy-primitives = "0.7"
 # TODO this is pulling in ed25519 crate unaccelerated, look into if actually used.
 tendermint = { version = "0.39.1", default-features = false }
 tendermint-proto = { version = "0.39.1", default-features = false }
